# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atiende Rentas Vacacionales — API
  version: 0.3.0
  description: |
    Contrato de API de `apps/api` (Lote 3: auth/roles/RLS/auditoría sobre el
    dominio de calendario de Lote 1). Fuente de verdad ejecutable: los
    esquemas `zod` de `apps/api/src/contrato/tipos.ts` (importables como
    `@atiende-rv/api/contrato`) — este archivo es la descripción
    human-readable del mismo contrato, publicada en el primer commit de
    este lote para que Lote 4 (frontend) pueda empezar contra tipos
    estables antes de que todos los endpoints tengan implementación
    completa.

    Reglas de dominio no negociables reflejadas en este contrato (D-006,
    D-011, BLUEPRINT): la API NUNCA expone cancelación ni modificación de
    una reserva cuyo origen sea un canal externo (Airbnb/Vrbo/Booking) —
    solo reservas DIRECTAS (creadas por esta misma API). Tampoco expone
    envío de mensajes a huéspedes (eso es Lote 6, con aprobación humana
    obligatoria). `POST /canales/:id/sync` solo ENCOLA una solicitud de
    sincronización manual en `outbox_evento`; la ejecución real la hace el
    worker de Lote 2.
servers:
  - url: http://localhost:8787
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /usuarios:
    get:
      tags: [usuarios]
      summary: Listar usuarios del tenant — admin_gestora/superadmin
      responses:
        "200": { description: Lista de usuarios }
        "403": { $ref: "#/components/responses/RolForbidden" }
    post:
      tags: [usuarios]
      summary: Alta de un colaborador/rol interno dentro del tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearUsuario" }
      responses:
        "201": { description: Usuario creado }
        "403": { $ref: "#/components/responses/RolForbidden" }
        "422": { $ref: "#/components/responses/Validacion" }
...
  /unidades:
    get:
      tags: [unidades]
      responses:
        "200": { description: Lista de unidades }
    post:
      tags: [unidades]
      summary: Crear unidad — admin_gestora/superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearUnidad" }
      responses:
        "201": { description: Unidad creada }
        "403": { $ref: "#/components/responses/RolForbidden" }
...
  /pricing/unidades/{unidadId}:
    get:
      tags: [pricing]
      summary: Contexto de pricing completo de una unidad (base, temporadas, descuentos, min-stay)
      parameters:
        - name: unidadId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Contexto de pricing }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
  /backoffice/unidades:
    get:
      tags: [backoffice]
      summary: CRUD de administración de unidades
      parameters:
        - name: propiedadId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Unidades de la propiedad }
    post:
      tags: [backoffice]
      summary: Alta de unidad(es) — `cantidad` crea N unidades independientes de una sola vez (H-012)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearUnidadBackoffice" }
      responses:
        "201": { description: Unidad(es) creada(s) }
  /backoffice/unidades/{id}:
    patch:
      tags: [backoffice]
      summary: Actualizar unidad
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Unidad actualizada }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    delete:
      tags: [backoffice]
      summary: Eliminar unidad
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Unidad eliminada }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
  /backoffice/usuarios:
    get:
      tags: [backoffice]
      summary: Usuarios/roles del tenant
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Usuarios del tenant }
  /backoffice/usuarios/invitaciones:
    get:
      tags: [backoffice]
      summary: Invitaciones del tenant
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Invitaciones del tenant }
    post:
      tags: [backoffice]
      summary: Invitar a un colaborador (token devuelto UNA sola vez, nunca persistido en claro)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearInvitacion" }
      responses:
        "201": { description: Invitación creada (incluye `token` una sola vez) }
  /backoffice/usuarios/invitaciones/{id}/revocar:
    post:
      tags: [backoffice]
      summary: Revocar una invitación pendiente
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Invitación revocada }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Validacion:
      description: Error de validación de entrada (zod)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
...
    RolForbidden:
      description: "rol_forbidden: el rol/nivel de colaborador no tiene permiso para esta acción"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
    NoEncontrado:
      description: "recurso_no_encontrado (incluye recursos de otro tenant, ocultos por RLS — nunca confirma su existencia)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
  schemas:
    ErrorHttp:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum:
                - credenciales_invalidas
                - token_invalido
                - token_expirado
                - tenant_forbidden
                - rol_forbidden
                - rango_invalido
                - unidad_no_disponible
                - conflicto_pendiente
                - reserva_no_directa
                - recurso_no_encontrado
                - validacion
                - rate_limited
                - error_interno
            mensaje: { type: string }
            detalles: {}
...
    RolUsuario:
      type: string
      enum: [superadmin, admin_gestora, operador, limpieza, propietario, contador]
    ColaboradorNivel:
      type: string
      nullable: true
      enum: [acceso_total, calendario_mensajeria, solo_calendario]
...
    CuerpoCrearUsuario:
      type: object
      required: [email, password, rol]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        rol: { $ref: "#/components/schemas/RolUsuario" }
        colaboradorNivel: { $ref: "#/components/schemas/ColaboradorNivel" }
        ownerId: { type: string, format: uuid }
...
    CuerpoCrearUnidad:
      type: object
      required: [propiedadId, nombre]
      properties:
        propiedadId: { type: string, format: uuid }
        nombre: { type: string }
        ownerId: { type: string, format: uuid }
        duracionMinimaNoches: { type: integer, minimum: 1 }
...
    CuerpoCrearUnidadBackoffice:
      type: object
      required: [propiedadId, nombre]
      properties:
        propiedadId: { type: string, format: uuid }
        nombre: { type: string }
        ownerId: { type: string, format: uuid }
        duracionMinimaNoches: { type: integer, minimum: 1 }
        cantidad: { type: integer, minimum: 1, maximum: 50, default: 1, description: "Crea N unidades independientes (H-012)" }
...
    CuerpoCrearInvitacion:
      type: object
      required: [email, rol]
      properties:
        email: { type: string, format: email }
        rol: { type: string, enum: [superadmin, admin_gestora, operador, limpieza, propietario, contador] }
        colaboradorNivel: { type: string, enum: [acceso_total, calendario_mensajeria, solo_calendario] }
        ownerId: { type: string, format: uuid }
        ttlHoras: { type: integer, minimum: 1, maximum: 720, default: 72 }